date time

+++ src/dwpbay/Item.java
@@ -11,10 +11,10 @@ public class Item {
 	private String picture ;
 	private Double price ;
 	private boolean onSale ;
-	private Double reservePrice ;
+	private Float reservePrice ;
 	private LocalDateTime startDateTime ;
 	
-	public Item(int itemID, String title, String category, String description, String picture, Double price, boolean onSale, Double reservePrice, LocalDateTime startDateTime) {
+	public Item(int itemID, String title, String category, String description, String picture, Double price, boolean onSale, Float reservePrice, LocalDateTime startDateTime) {
 		this.itemID = itemID ;
 		this.title = title ;
 		this.category = category ;
